requirements.txt
chromadb==0.4.24
langchain==0.1.12
langchain-groq==0.0.7
langchain_community==0.0.25
langchain_core==0.1.30
python-dotenv==1.0.1
pandas==2.2.1
streamlit==1.31.1
transformers==4.38.2
torch==2.2.1
matplotlib==3.8.3
yfinance==0.2.37
alpha-vantage==2.3.1
TA-Lib==0.4.28
sentence-transformers==2.5.1
beautifulsoup4==4.12.3
requests==2.31.0
alpha_vantage